Toolbar-Buttons more standard
1) The "Cancel" button maybe would be placed more standard to the left
not center (take a look at the same button in the camera-plugin); also
such will be a (again more suitable) place for the coming flash toggle.

2) The button for flipping cameras has the wrong icon (again: take a
look at the camera plugin); the new icon is taken from the icons8.com
collection, which should be free for Open Source, else take a look at
the accompanying license).