orientation settings independent to the calling App
Since commit
https://github.com/conveyGmbH/phonegap-plugin-barcodescanner/commit/9d18
4d509b2596a5d71ea8af8b600e4555660386 the plugin always supported all
possible orientations of the calling App. Now these can be overridden,
e.g., if the App only supports Portrait, the scanner could nonetheless
also support landscape (for better recognition of long one-dimensional
codes).

The orientation-dependent setting of the view- and preview-layer-frames
was a) wrong placed in loadView (here the orientation is still that of
the parent) instead of viewWillAppear; and b) wrongly simply took the
parent-view’s frame (in it’s orientation) – instead of explicitly
realign it to a possibly other orientation.